[cleanup] Move `wxcApp` in its own file. Rename main.cpp into mainFrame.cpp
